Move SessionAuthenticationHelper to pyramid.authentication.
